souse/requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
alchemy==20.5
alembic==1.13.0
annotated-types==0.6.0
async-timeout==4.0.3
asyncpg==0.29.0
attrs==23.1.0
certifi==2023.11.17
chardet==3.0.4
charset-normalizer==3.3.2
docutils==0.20.1
filelock==3.0.12
frozenlist==1.4.0
greenlet==3.0.1
idna==2.8
lockfile==0.12.2
Mako==1.3.0
MarkupSafe==2.1.3
multidict==6.0.4
psycopg2-binary==2.9.9
pydantic==2.5.2
pydantic_core==2.14.5
python-daemon==2.2.4
python-dotenv==1.0.0
request2==0.2
requests==2.22.0
SQLAlchemy==2.0.23
typing_extensions==4.8.0
urllib3==1.25.11
yarl==1.9.3
